feat(suite): `bismark` whole-suite meta-crate — `cargo install bismark` installs all 12 tools (install-story W1)
Adds a `bismark` meta / batteries-included crate so `cargo install bismark`
builds the entire suite (all 12 canonical binaries) in one command — the
user-facing install premise + claims the bare `bismark` crate name on crates.io.

- New `rust/bismark` crate (v2.0.0): 12 `src/bin/<tool>.rs` wrappers, each a
  thin `fn main() -> ExitCode { <tool_crate>::run_main() }` + the shared mimalloc
  `#[global_allocator]` (so its binaries match the standalone tools exactly).
  Exact-pinned to the coordinated GA suite versions (=1.0.0 / =1.2.1).
- Each of the 12 tool crates now exposes `pub fn run_main() -> ExitCode` in its
  lib (the former `main.rs` body); `main.rs` is a thin wrapper. Byte-inert
  (pure entry-point plumbing) — per-crate quirks preserved (aligner @PG argv
  capture, `--man` guards, the filter-nonconversion empty-files gate + bare-`{e}`
  prints, summary's txt-then-html order). perl-oracle + full workspace tests green.
- `default-members` excludes `bismark` so the release tarball + plain `cargo build`
  (neither passes --workspace) build only the tool crates — collision-free (the
  meta-crate's bins share the canonical names). `cargo build -p bismark` /
  `cargo install bismark` build it explicitly.
- Docs (W2): root README → Rust-first Installation (`cargo install bismark`,
  container, prebuilt binaries) + reconciled GA banner + Perl-legacy section;
  rust/README de-staled (`_rs`→canonical, `:beta`→`:2.0.0`/`:latest`,
  `cargo install --git …`→`cargo install bismark`, no-samtools, bioconda-pending note).

Verified: `cargo install --path bismark` installs all 12 bins (each runs, uniform
--version + footer); fmt + clippy(-D warnings) clean; `cargo test --workspace`
passes; `cargo publish --workspace --dry-run` 15/15 (meta-crate packages).

Follow-ups: publish `bismark 2.0.0` to crates.io (separate step — the GA tag
already exists, so a targeted publish); docs-site `_rs` sweep; bioconda (W3).
