+++ README.md
@@ -3,13 +3,14 @@
 [![install with bioconda](https://img.shields.io/badge/install%20with-bioconda-brightgreen.svg?style=flat)](http://bioconda.github.io/recipes/bismark/README.html)
 
 > [!IMPORTANT]
-> **The Perl version of Bismark is in maintenance freeze.** Bismark `v0.25.x` (this codebase) now
-> receives **critical correctness and security fixes only** — no new features or performance changes.
-> It is being succeeded by the **[Bismark Rust suite](https://felixkrueger.github.io/Bismark/rust/overview/)**,
-> which is byte-identical to `v0.25.1` on the faithful default path, faster, lower-memory, and actively
-> developed (currently in beta); the Perl code will be archived as tagged legacy at the Rust general
-> release. **New contributions, including performance work, should target the Rust suite** — please read
-> [CONTRIBUTING.md](CONTRIBUTING.md) before opening a pull request. Thank you!
+> **Bismark is now the Rust suite — v2.0.0 is generally available.** The bisulfite aligner and
+> methylation tools have been rewritten from Perl to Rust: **byte-identical** to Perl `v0.25.1` on the
+> faithful default path, faster, and lower-memory — and this is the **supported default**. Get it via
+> [Installation](#installation) (`cargo install bismark`, container, or prebuilt binaries) and see the
+> **[Rust suite overview](https://felixkrueger.github.io/Bismark/rust/overview/)**.
+> The original **Perl `v0.25.x`** (the scripts at this repo root) is now **legacy / maintenance-freeze**
+> (critical fixes only; tagged [`v0.25.1`](https://github.com/FelixKrueger/Bismark/releases/tag/v0.25.1)).
+> **New contributions should target the Rust suite** — see [CONTRIBUTING.md](CONTRIBUTING.md).
 
 > **See the documentation**: <https://felixkrueger.github.io/Bismark>
 
@@ -21,8 +22,6 @@ Bismark is a program to map bisulfite treated sequencing reads to a genome of in
 - Alignment seed length, number of mismatches etc. are adjustable
 - Output discriminates between cytosine methylation in `CpG`, `CHG` and `CHH` context
 
-> **Rust rewrite in progress:** Bismark's post-alignment tools are being ported from Perl to Rust — faster, with byte-identical output. See [`rust/`](rust/) for the workspace and a per-tool status table, and [Installing](rust/README.md#installing) to get the Rust suite (prebuilt binaries, container, or `cargo install`).
-
 ## Documentation
 
 The Bismark documentation can be found with the code in the [docs](docs) subfolder and can also be read online: <https://felixkrueger.github.io/Bismark/>
@@ -31,12 +30,26 @@ There is also an overview of the alignment modes that are currently supported by
 
 ## Installation
 
-Bismark is written in Perl and is executed from the command line. To install Bismark simply download the latest release of the code from the [Releases page](https://github.com/FelixKrueger/Bismark/releases) and extract the files into a Bismark installation folder.
+Bismark **v2.0.0+** is the Rust suite. Pick one:
 
-Bismark needs the following tools to be installed and ideally available in the `PATH` environment:
+```bash
+# 1. crates.io — installs the whole suite (all tools) into ~/.cargo/bin
+cargo install bismark
 
-- [Bowtie2](http://bowtie-bio.sourceforge.net/bowtie2/) or [HISAT2](https://ccb.jhu.edu/software/hisat2/index.shtml) or [minimap2](https://lh3.github.io/minimap2/minimap2.html)
-- [Samtools](http://www.htslib.org/)
+# 2. Container image (nothing to install; drop-in for nf-core/methylseq)
+docker pull ghcr.io/felixkrueger/bismark:2.0.0     # or :latest
+
+# 3. Prebuilt binaries — download from the Releases page and put on your PATH
+#    https://github.com/FelixKrueger/Bismark/releases
+```
+
+**External tools on your `PATH`:** an aligner — [Bowtie2](http://bowtie-bio.sourceforge.net/bowtie2/) (default), or optionally [HISAT2](https://ccb.jhu.edu/software/hisat2/index.shtml) or [minimap2](https://lh3.github.io/minimap2/minimap2.html). **No Samtools needed** — all BAM/SAM/CRAM I/O is pure-Rust. The container bundles the aligners for you. See [`rust/README.md`](rust/README.md#installing) for details and per-tool installs.
+
+> A bioconda `mamba install bismark` for the Rust suite is coming as a fast-follow; until then the bioconda `bismark` package is the legacy Perl `v0.25.x`.
+
+### Legacy: the Perl Bismark (v0.25.x)
+
+The original Perl scripts remain at this repo root (maintenance-freeze). To use them, download the [`v0.25.1` release](https://github.com/FelixKrueger/Bismark/releases/tag/v0.25.1) (or `mamba install bismark=0.25.1`); they need [Bowtie2](http://bowtie-bio.sourceforge.net/bowtie2/)/[HISAT2](https://ccb.jhu.edu/software/hisat2/index.shtml)/[minimap2](https://lh3.github.io/minimap2/minimap2.html) **and** [Samtools](http://www.htslib.org/) on the `PATH`.
 
 ## Links
