+++ rust/Cargo.toml
@@ -1,6 +1,14 @@
 [workspace]
 resolver = "2"
-members = ["bismark-meta", "bismark-io", "bismark-dedup", "bismark-extractor", "bismark-methylation-consistency", "bismark-bedgraph", "bismark-coverage2cytosine", "bismark-genome-preparation", "bismark-nome-filtering", "bismark-bam2nuc", "bismark-filter-nonconversion", "bismark-report", "bismark-summary", "bismark-aligner"]
+members = ["bismark-meta", "bismark-io", "bismark-dedup", "bismark-extractor", "bismark-methylation-consistency", "bismark-bedgraph", "bismark-coverage2cytosine", "bismark-genome-preparation", "bismark-nome-filtering", "bismark-bam2nuc", "bismark-filter-nonconversion", "bismark-report", "bismark-summary", "bismark-aligner", "bismark"]
+# The `bismark` meta-crate re-declares every tool's canonical `[[bin]]` (so
+# `cargo install bismark` builds the whole suite). Those bin names collide with
+# the individual tool crates' bins, so exclude it from default builds — release
+# tarball + plain `cargo build` (neither passes --workspace) build only the tool
+# crates, collision-free. Build/install the meta-crate explicitly: `cargo build
+# -p bismark` / `cargo install bismark`. (An explicit `--workspace` still builds
+# it and warns about the collisions — cosmetic; CI stays green.)
+default-members = ["bismark-meta", "bismark-io", "bismark-dedup", "bismark-extractor", "bismark-methylation-consistency", "bismark-bedgraph", "bismark-coverage2cytosine", "bismark-genome-preparation", "bismark-nome-filtering", "bismark-bam2nuc", "bismark-filter-nonconversion", "bismark-report", "bismark-summary", "bismark-aligner"]
 
 [workspace.package]
 edition = "2024"
