+++ rust/bismark-bam2nuc/src/main.rs
@@ -1,44 +1,19 @@
-//! Binary entry point for `bam2nuc`.
+//! Binary entry point for `bam2nuc` — thin wrapper over
+//! [`bismark_bam2nuc::run_main`] (shared with the `bismark` meta-crate's bin so
+//! `cargo install bismark` and `cargo install bismark-bam2nuc` behave identically).
 //!
-//! Parses [`Cli`], handles `--version`, validates into a [`ResolvedConfig`],
-//! then runs the nucleotide-coverage report.
-//!
-//! Exit codes: `0` success · `1` any [`BismarkBam2nucError`] · `2` clap parse
-//! error (clap convention, emitted by `Cli::parse`).
+//! Exit codes: `0` success · `1` any [`bismark_bam2nuc::BismarkBam2nucError`] ·
+//! `2` clap parse error.
 
 use std::process::ExitCode;
 
-use clap::Parser;
-
-use bismark_bam2nuc::cli::Cli;
-use bismark_bam2nuc::error::BismarkBam2nucError;
-use bismark_bam2nuc::version_string;
-
 // Multithreaded allocator (#884/#915 sibling precedent). Allocator-only — the
 // per-read counting loop allocates a span Vec per read; mimalloc trims the
-// malloc cost. Output is byte-identical.
+// malloc cost. Output is byte-identical. Kept in the binary crate root (each
+// binary — this one and the meta-crate's — sets its own).
 #[global_allocator]
 static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
 
 fn main() -> ExitCode {
-    let cli = Cli::parse();
-
-    // `--version` / `-V` handled here (clap auto-version disabled in cli.rs).
-    if cli.version {
-        println!("{}", version_string());
-        return ExitCode::SUCCESS;
-    }
-
-    match run(cli) {
-        Ok(()) => ExitCode::SUCCESS,
-        Err(e) => {
-            eprintln!("error: {e}");
-            ExitCode::from(1)
-        }
-    }
-}
-
-fn run(cli: Cli) -> Result<(), BismarkBam2nucError> {
-    let config = cli.validate()?;
-    bismark_bam2nuc::run(&config)
+    bismark_bam2nuc::run_main()
 }
