+++ rust/bismark-methylation-consistency/src/main.rs
@@ -1,42 +1,15 @@
-//! Binary entry point for `methylation_consistency`.
-//!
-//! Parses CLI via [`bismark_methylation_consistency::Cli`], validates into a
-//! [`ResolvedConfig`], then runs the per-file consistency split via
-//! [`pipeline::run`].
+//! Binary entry point for `methylation_consistency` — thin wrapper over
+//! [`bismark_methylation_consistency::run_main`] (shared with the `bismark`
+//! meta-crate's bin so `cargo install bismark` and `cargo install
+//! bismark-methylation-consistency` behave identically).
 //!
 //! Exit codes:
 //! - `0` — success
-//! - `1` — any [`MethConsError`]
+//! - `1` — any [`bismark_methylation_consistency::error::MethConsError`]
 //! - `2` — clap parse error (clap convention for usage errors)
 
 use std::process::ExitCode;
 
-use clap::Parser;
-
-use bismark_methylation_consistency::cli::Cli;
-use bismark_methylation_consistency::error::MethConsError;
-use bismark_methylation_consistency::pipeline;
-use bismark_methylation_consistency::version_string;
-
 fn main() -> ExitCode {
-    let cli = Cli::parse();
-
-    // `--version` / `-V` handled here (clap auto-version disabled in cli.rs).
-    if cli.version {
-        println!("{}", version_string());
-        return ExitCode::SUCCESS;
-    }
-
-    match run(cli) {
-        Ok(()) => ExitCode::SUCCESS,
-        Err(e) => {
-            eprintln!("error: {e}");
-            ExitCode::from(1)
-        }
-    }
-}
-
-fn run(cli: Cli) -> Result<(), MethConsError> {
-    let config = cli.validate()?;
-    pipeline::run(&config)
+    bismark_methylation_consistency::run_main()
 }
