+++ rust/bismark/src/bin/bismark_methylation_extractor.rs
@@ -1,0 +1,11 @@
+//! `bismark_methylation_extractor` — one of the binaries installed by `cargo install bismark`.
+//! Thin wrapper over [`bismark_extractor::run_main`]; byte-identical to that tool's own
+//! binary (each sets the same multithreaded allocator).
+use std::process::ExitCode;
+
+#[global_allocator]
+static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
+
+fn main() -> ExitCode {
+    bismark_extractor::run_main()
+}
