+++ rust/bismark/src/bin/filter_non_conversion.rs
@@ -1,0 +1,11 @@
+//! `filter_non_conversion` — one of the binaries installed by `cargo install bismark`.
+//! Thin wrapper over [`bismark_filter_nonconversion::run_main`]; byte-identical to that tool's own
+//! binary (each sets the same multithreaded allocator).
+use std::process::ExitCode;
+
+#[global_allocator]
+static GLOBAL: mimalloc::MiMalloc = mimalloc::MiMalloc;
+
+fn main() -> ExitCode {
+    bismark_filter_nonconversion::run_main()
+}
